requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 25.3.1
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
aiohappyeyeballs=2.4.4=py311h06a4308_0
aiohttp=3.11.10=py311h5eee18b_0
aiosignal=1.2.0=pyhd3eb1b0_0
# anytree=2.12.1=pypi_0
arrow-cpp=19.0.0=h865e1df_2
attrs=24.3.0=py311h06a4308_0
aws-c-auth=0.6.19=h5eee18b_0
aws-c-cal=0.5.20=hdbd6064_0
aws-c-common=0.8.5=h5eee18b_0
aws-c-compression=0.2.16=h5eee18b_0
aws-c-event-stream=0.2.15=h6a678d5_0
aws-c-http=0.6.25=h5eee18b_0
aws-c-io=0.13.10=h5eee18b_0
aws-c-mqtt=0.7.13=h5eee18b_0
aws-c-s3=0.1.51=hdbd6064_0
aws-c-sdkutils=0.1.6=h5eee18b_0
aws-checksums=0.1.13=h5eee18b_0
aws-crt-cpp=0.18.16=h6a678d5_0
aws-sdk-cpp=1.11.212=hecad206_0
# better-mistakes=1.0=pypi_0
blas=1.0=mkl
boost-cpp=1.82.0=hdb19cb5_2
bottleneck=1.4.2=py311hf4808d0_0
brotli-python=1.0.9=py311h6a678d5_9
bzip2=1.0.8=h5eee18b_6
c-ares=1.19.1=h5eee18b_0
ca-certificates=2025.2.25=h06a4308_0
certifi=2025.1.31=py311h06a4308_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
# clean=0.1=dev_0
click=8.1.7=py311h06a4308_0
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
datasets=3.3.2=py311h06a4308_0
dill=0.3.8=py311h06a4308_0
# einop=0.0.1=pypi_0
# einops=0.8.1=pypi_0
# fair-esm=2.0.0=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py311h06a4308_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.5.0=py311h5eee18b_0
fsspec=2024.12.0=py311h06a4308_0
gflags=2.2.2=h6a678d5_1
giflib=5.2.2=h5eee18b_0
glog=0.5.0=h6a678d5_1
gmp=6.3.0=h6a678d5_0
gmpy2=2.2.1=py311h5eee18b_0
gnutls=3.6.15=he1e5248_0
# hierarchicalsoftmax=1.1.0=pypi_0
huggingface_hub=0.29.2=py311h06a4308_0
icu=73.1=h6a678d5_0
idna=3.7=py311h06a4308_0
intel-openmp=2023.1.0=hdb19cb5_46306
jinja2=3.1.6=py311h06a4308_0
joblib=1.4.2=py311h06a4308_0
jpeg=9e=h5eee18b_3
krb5=1.20.1=h143b758_1
lame=3.100=h7b6447c_0
lcms2=2.16=hb9589c4_0
ld_impl_linux-64=2.40=h12ee557_0
lerc=4.0.0=h6a678d5_0
libabseil=20250127.0=cxx17_h6a678d5_0
libboost=1.82.0=h109eef0_2
libbrotlicommon=1.0.9=h5eee18b_9
libbrotlidec=1.0.9=h5eee18b_9
libbrotlienc=1.0.9=h5eee18b_9
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.9.1.3=0
libcurand=10.3.5.147=0
libcurl=8.12.1=hc9e6f67_0
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libdeflate=1.22=h5eee18b_0
libedit=3.1.20230828=h5eee18b_0
libev=4.33=h7f8727e_1
libevent=2.1.12=hdbd6064_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libgrpc=1.71.0=h2d74bed_0
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libnghttp2=1.57.0=h2d74bed_0
libnpp=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libpng=1.6.39=h5eee18b_0
libprotobuf=5.29.3=hc99497a_0
libre2-11=2024.07.02=h6a678d5_0
libssh2=1.11.1=h251f7ec_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libthrift=0.15.0=h1795dd8_2
libtiff=4.5.1=hffd6297_1
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=h5eee18b_1
llvm-openmp=14.0.6=h9e868ea_0
lz4-c=1.9.4=h6a678d5_1
# markdown-it-py=3.0.0=pypi_0
markupsafe=3.0.2=py311h5eee18b_0
# mdurl=0.1.2=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py311h5eee18b_2
mkl_fft=1.3.11=py311h5eee18b_0
mkl_random=1.2.8=py311ha02d727_0
mpc=1.3.1=h5eee18b_0
mpfr=4.2.1=h5eee18b_0
mpmath=1.3.0=py311h06a4308_0
multidict=6.1.0=py311h5eee18b_0
multiprocess=0.70.15=py311h06a4308_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.4.2=py311h06a4308_0
nltk=3.9.1=py311h06a4308_0
numexpr=2.10.1=py311h3c60e43_0
numpy=1.26.4=py311h08b1b3b_0
numpy-base=1.26.4=py311hf175353_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.5.2=he7f1fd0_0
openssl=3.0.16=h5eee18b_0
orc=2.1.1=hd396ef6_0
packaging=24.2=py311h06a4308_0
pandas=2.2.3=py311h6a678d5_0
pillow=11.1.0=py311hcea889d_0
pip=25.0=py311h06a4308_0
propcache=0.2.0=py311h5eee18b_0
protobuf=5.29.3=py311he621ea3_0
pyarrow=19.0.0=py311h6a678d5_1
# pygments=2.19.1=pypi_0
pysocks=1.7.1=py311h06a4308_0
python=3.11.11=he870216_0
python-dateutil=2.9.0post0=py311h06a4308_2
python-tzdata=2023.3=pyhd3eb1b0_0
python-xxhash=3.5.0=py311h5eee18b_0
pytorch=2.5.1=py3.11_cuda11.8_cudnn9.1.0_0
pytorch-cuda=11.8=h7e8668a_6
pytorch-mutex=1.0=cuda
pytz=2024.1=py311h06a4308_0
pyyaml=6.0.2=py311h5eee18b_0
re2=2024.07.02=hdb19cb5_0
readline=8.2=h5eee18b_0
regex=2024.11.6=py311h5eee18b_0
requests=2.32.3=py311h06a4308_1
# rich=14.0.0=pypi_0
s2n=1.3.27=hdbd6064_0
safetensors=0.5.3=py311h448de89_0
scikit-learn=1.6.1=py311h6a678d5_0
scipy=1.15.2=py311h23a989f_1
setuptools=75.8.0=py311h06a4308_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.2.1=h6a678d5_0
sqlite=3.45.3=h5eee18b_0
# sympy=1.13.1=pypi_0
tbb=2021.8.0=hdb19cb5_0
threadpoolctl=3.5.0=py311h92b7b1e_0
tk=8.6.14=h39e8969_0
tokenizers=0.21.0=py311h5e3497e_0
torchaudio=2.5.1=py311_cu118
torchtriton=3.1.0=py311
torchvision=0.20.1=py311_cu118
tqdm=4.67.1=py311h92b7b1e_0
transformers=4.49.0=py311h06a4308_0
typing-extensions=4.12.2=py311h06a4308_0
typing_extensions=4.12.2=py311h06a4308_0
tzdata=2025a=h04d1e81_0
urllib3=2.3.0=py311h06a4308_0
utf8proc=2.6.1=h5eee18b_1
wheel=0.45.1=py311h06a4308_0
xxhash=0.8.0=h7f8727e_3
xz=5.6.4=h5eee18b_1
yaml=0.2.5=h7b6447c_0
yarl=1.18.0=py311h5eee18b_0
zip=3.0=h7f8727e_1
zlib=1.2.13=h5eee18b_1
zstd=1.5.6=hc292b87_0
